Refactoring: update topic and messages structure according to docs

+++ lib/Core/Date/Date.h
@@ -51,7 +51,7 @@ class Date : public ICoreEventBusKitConnection, public ISchedulerKitConnection, 
   void attach() override {}
 
   bool store() override {
-    object().put("epoch", mNTPClient.getEpochTime());
+    object().put("epoch", (long) mNTPClient.getEpochTime());
     return CBORStorage::store();
   }
